requirements.txt
asgiref==3.7.2
Django==5.0
django-admin-rangefilter==0.12.0
django-cors-headers==4.3.1
imageio==2.33.0
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
scikit-image==0.22.0
scipy==1.11.4
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
tifffile==2023.9.26
tzdata==2023.3
